1. Change hello_seq.c to be pageinfo.c.
2. Change Makefile to include the new module.

3. Implement another module to export invidual page content.

This module will create an entry named as singlepage in /proc. To get the
content of the physical page with id n, write the number n into
/proc/singlepage and then cat /proc/singlepage.

e.g.
	$ echo 100000 > /proc/singlepage
	$ cat /proc/singlepage

TODO: The output of the content of each page is interpretted as a string, which
means it is really hard for us to see what data is there. Ideally, we want to
see hexadecimal numbers.